+++ CHANGELOG.md
@@ -1,16 +1,26 @@
+## Unreleased
+
+**[compare](https://github.com/metosin/komponentit/compare/0.2.2...master)**
+
 ## 0.2.2
 
 - Removed `console.log` from autosize input
 
+**[compare](https://github.com/metosin/komponentit/compare/0.2.1...0.2.2)**
+
 ## 0.2.1
 
 - Reimplemented autosize input
+
+**[compare](https://github.com/metosin/komponentit/compare/0.2.0...0.2.1)**
 
 ## 0.2.0
 
 - New autosize input element
 - Refactored autocomplete styles etc.?
 - Lots of stuff
+
+**[compare](https://github.com/metosin/komponentit/compare/0.1.3...0.2.0)**
 
 ## 0.1.3 (13.6.2016)